+++ app/src/ads/java/app/opendocument/droid/nonfree/Linked.kt
@@ -2,3 +2,6 @@ package app.opendocument.droid.nonfree
 
 /** Read through [Features]. */
 internal const val LINKS_ADS = true
+
+/** Read through [Features]. */
+internal const val ADVANCED_EDITING = false